@@ -3,39 +3,39 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "ng serve --proxy-config proxy.conf.js",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.0.0-next.0",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "~13.0.0-next.0",
-    "@angular/compiler": "~13.0.0-next.0",
-    "@angular/core": "~13.0.0-next.0",
-    "@angular/forms": "~13.0.0-next.0",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "~13.0.0-next.0",
-    "@angular/platform-browser-dynamic": "~13.0.0-next.0",
-    "@angular/router": "~13.0.0-next.0",
-    "rxjs": "~7.4.0",
+    "@angular/animations": "^17.0.7",
+    "@angular/cdk": "^17.0.4",
+    "@angular/common": "^17.0.7",
+    "@angular/compiler": "^17.0.7",
+    "@angular/core": "^17.0.7",
+    "@angular/forms": "^17.0.7",
+    "@angular/material": "^17.0.4",
+    "@angular/platform-browser": "^17.0.7",
+    "@angular/platform-browser-dynamic": "^17.0.7",
+    "@angular/router": "^17.0.7",
+    "rxjs": "~6.6.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.14.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.0.0",
-    "@angular/cli": "~13.0.0",
-    "@angular/compiler-cli": "~13.0.0-next.0",
-    "@types/jasmine": "~3.10.0",
+    "@angular-devkit/build-angular": "^17.0.7",
+    "@angular/cli": "^17.0.7",
+    "@angular/compiler-cli": "^17.0.7",
+    "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
-    "jasmine-core": "~3.10.0",
+    "jasmine-core": "~3.8.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~5.2.2"
   }
 }
