@@ -1,11 +1,11 @@
-import { Component, NgModule } from '@angular/core';
-import {MatTableModule, MatTableDataSource} from '@angular/material/table';
-
+import { Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
+import { MatToolbarModule } from '@angular/material/toolbar';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {
   title = 'crud-angular';
