@@ -8,8 +8,8 @@ describe('ErrorDialogComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ ErrorDialogComponent ]
-    })
+    imports: [ErrorDialogComponent]
+})
     .compileComponents();
   });
 
